Extract node_storage as well
